+++ example-model/src/main/java/com/appunite/rx/example/dao/auth/MyCurrentLoggedInUserDao.java
@@ -1,0 +1,7 @@
+package com.appunite.rx.example.dao.auth;
+
+import com.appunite.login.CurrentLoggedInUserDao;
+
+public interface MyCurrentLoggedInUserDao extends CurrentLoggedInUserDao {
+
+}
