--- example-model/src/main/java/com/appunite/rx/example/model/dao/MyCurrentLoggedInUserDao.java
@@ -1,7 +1,0 @@
-package com.appunite.rx.example.model.dao;
-
-import com.appunite.login.CurrentLoggedInUserDao;
-
-public interface MyCurrentLoggedInUserDao extends CurrentLoggedInUserDao {
-
-}
