--- example-model/src/main/java/com/appunite/rx/example/model/model/Post.java
@@ -1,20 +1,0 @@
-package com.appunite.rx.example.model.model;
-
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
-
-public class Post extends PostId {
-    @Nonnull
-    private final String name;
-
-    public Post(@Nonnull String id,
-                @Nonnull String name) {
-        super(id);
-        this.name = name;
-    }
-
-    @Nonnull
-    public String name() {
-        return name;
-    }
-}
