--- example-model/src/main/java/com/appunite/rx/example/model/model/PostId.java
@@ -1,18 +1,0 @@
-package com.appunite.rx.example.model.model;
-
-import javax.annotation.Nonnull;
-
-public class PostId {
-    @Nonnull
-    private final String id;
-
-    public PostId(@Nonnull String id) {
-        this.id = id;
-    }
-
-    @Nonnull
-    public String id() {
-        return id;
-    }
-
-}
