--- example/src/main/java/com/appunite/rx/example/BaseActivity.java
@@ -1,6 +1,0 @@
-package com.appunite.rx.example;
-
-import android.support.v7.app.AppCompatActivity;
-
-public class BaseActivity extends AppCompatActivity {
-}
