--- example/src/main/java/com/appunite/rx/example/MainApplication.java
@@ -1,6 +1,0 @@
-package com.appunite.rx.example;
-
-import android.app.Application;
-
-public class MainApplication extends Application {
-}
