+++ example/src/main/java/com/appunite/rx/example/application/MainApplication.java
@@ -1,0 +1,6 @@
+package com.appunite.rx.example.application;
+
+import android.app.Application;
+
+public class MainApplication extends Application {
+}
