+++ example/src/main/java/com/appunite/rx/example/ui/internal/BaseActivity.java
@@ -1,0 +1,6 @@
+package com.appunite.rx.example.ui.internal;
+
+import android.support.v7.app.AppCompatActivity;
+
+public class BaseActivity extends AppCompatActivity {
+}
